refactor(RichComposerInput): simplify caret handling and remove unused code
