+++ cometline/src/lib/components/RichComposerInput.svelte
@@ -1,10 +1,15 @@
 <script lang="ts">
-	import { onDestroy, tick } from 'svelte';
+	import { tick } from 'svelte';
 	import type { CaretTrailSettings } from '$lib/types';
+	import {
+		customCaret,
+		resetCustomCaret,
+		scheduleCustomCaretMeasure,
+		type CustomCaretState
+	} from '$lib/dom/custom-caret';
 	import { faviconUrl, domainFromUrl, isHttpUrl } from '$lib/markdown/embed';
 	import { openLink } from '$lib/open-link';
 	import { openWorkspaceFilePreview } from '$lib/workspace/open-file-preview';
-	import { viewportDeltaToLocal } from '$lib/dom/caret-geometry';
 
 	let {
 		value = $bindable(''),
@@ -32,78 +37,19 @@
 
 	let wrap = $state<HTMLDivElement | null>(null);
 	let editor = $state<HTMLDivElement | null>(null);
-	let customCaret = $state<HTMLSpanElement | null>(null);
+	let caretEl = $state<HTMLSpanElement | null>(null);
 	let trailPoly = $state<SVGPolygonElement | null>(null);
 	let focused = $state(false);
+	let caretReady = $state(false);
 	// Guard so our own DOM writes don't recursively re-trigger input handling.
 	let syncing = false;
 	// IME composition guard — Enter during candidate selection must not trigger send.
 	let composing = false;
-	let raf = 0;
-	// Guard so the temporary marker we insert to measure an empty line doesn't
-	// re-enter measurement via selectionchange.
-	let measuring = false;
-	let caretReady = $state(false);
 	// Mention state used by the parent composer to show/hide the file picker.
 	let lastMentionActive = $state(false);
 	let lastMentionQuery = $state('');
-	// Caret geometry in wrap-local coordinates.
-	let caretW = 2;
-	let caretH = 22.5;
-	// Trail smear model (mirrors cursor_tail.glsl): a quad is drawn between the
-	// previous caret position (tail) and the current target (head); both ends
-	// ease independently with easeOutCirc so the trail extends then collapses.
-	let originX = 0; // tail anchor (where the smear starts from)
-	let originY = 0;
-	let targetX = 0; // head target (where the caret is heading)
-	let targetY = 0;
-	let visualX = 0; // current rendered caret position (may lag target during animation)
-	let visualY = 0;
-	let animStart = 0; // performance.now() when the current move began
-	let animating = false;
-	// When true, head snaps to target immediately and only the trail quad animates.
-	let trailOnly = false;
-	// Timestamp of the last user text input — used to detect typing/IME moves.
-	let lastInputAt = 0;
 
 	let caretTrailEnabled = $derived(caretTrail.enabled);
-	let baseTrailOpacity = $derived(0.32 + clampUnit(caretTrail.intensity) * 0.5);
-	let caretStyle = $derived(`--rce-caret-color: ${caretColor}`);
-
-	// easeOutCirc — matches the shader's chosen easing curve.
-	function easeOutCirc(x: number): number {
-		const c = clampUnit(x);
-		return Math.sqrt(1 - (c - 1) * (c - 1));
-	}
-
-	// Move animation duration in ms, shorter when "speed" is high.
-	function moveDuration(): number {
-		return 90 + (1 - clampUnit(caretTrail.speed)) * 220;
-	}
-
-	// Short smear for same-line typing / IME composition.
-	function typingTrailDuration(): number {
-		return 90 + (1 - clampUnit(caretTrail.speed)) * 110;
-	}
-
-	function isTypingMove(dy: number): boolean {
-		return (composing || performance.now() - lastInputAt < 120) && !isLineCrossing(dy);
-	}
-
-	// Vertical jump (in px) above which a move is treated as a teleport: the
-	// caret snaps without a trail. We gate on vertical distance only — a one-line
-	// word-wrap moves the caret a long way horizontally but only ~1 line down,
-	// and should still animate; a click far away or programmatic jump spans
-	// multiple lines and should snap clean.
-	function maxTrailVerticalJump(): number {
-		return caretH * 1.5;
-	}
-
-	// A move whose vertical delta exceeds ~half a line is a line crossing
-	// (newline, word-wrap, selection extending down/up a line).
-	function isLineCrossing(dy: number): boolean {
-		return Math.abs(dy) > caretH * 0.5;
-	}
 
 	/**
 	 * Serializes the contenteditable DOM back to plain text. URL chips serialize
@@ -141,252 +87,17 @@
 		value = serialize(editor);
 	}
 
-	function clampUnit(value: number): number {
-		if (!Number.isFinite(value)) return 0;
-		return Math.min(1, Math.max(0, value));
-	}
-
-	function setCaretVisual(x: number, y: number) {
-		visualX = x;
-		visualY = y;
-		if (!customCaret) return;
-		customCaret.style.transform = `translate3d(${x}px, ${y}px, 0)`;
-	}
-
-	function snapCaretTo(x: number, y: number) {
-		if (raf) {
-			cancelAnimationFrame(raf);
-			raf = 0;
-		}
-		animating = false;
-		trailOnly = false;
-		targetX = originX = x;
-		targetY = originY = y;
-		setCaretVisual(x, y);
-		clearTrail();
-	}
-
-	function clearTrail() {
-		trailPoly?.setAttribute('points', '');
-	}
-
-	/**
-	 * Measures the caret position in wrap-local coords. For a collapsed caret
-	 * this is the caret itself; for a selection it's the *focus* end (the moving
-	 * end driven by Shift+Arrow or mouse drag), so the comet rides the part of
-	 * the selection the user is actively extending. The native blue highlight
-	 * stays underneath. On an empty line getClientRects() returns nothing, so we
-	 * temporarily insert a zero-width marker, measure it, then remove it.
-	 * Returns null if the selection isn't inside the editor.
-	 */
-	function readCaretRect(): { x: number; y: number; h: number } | null {
-		if (!wrap || !editor) return null;
-		const selection = window.getSelection();
-		if (!selection || selection.rangeCount === 0 || selection.focusNode == null) return null;
-		const focusNode = selection.focusNode;
-		if (focusNode !== editor && !editor.contains(focusNode)) return null;
-
-		// Build a collapsed range at the selection's focus (moving) end.
-		const range = document.createRange();
-		try {
-			range.setStart(focusNode, selection.focusOffset);
-		} catch {
-			return null;
-		}
-		range.collapse(true);
-
-		const lineHeight = Number.parseFloat(getComputedStyle(editor).lineHeight) || 22.5;
-
-		let rect: DOMRect | undefined = range.getClientRects()[0];
-		if (!rect || (rect.width === 0 && rect.height === 0)) {
-			// Empty line / boundary: insert a zero-width marker to get a real rect.
-			// We snapshot the live selection, probe, then restore it exactly so a
-			// drag selection isn't collapsed by our measurement.
-			measuring = true;
-			const snap = {
-				anchorNode: selection.anchorNode,
-				anchorOffset: selection.anchorOffset,
-				focusNode: selection.focusNode,
-				focusOffset: selection.focusOffset
-			};
-			const marker = document.createElement('span');
-			marker.textContent = '\u200b';
-			const probe = range.cloneRange();
-			probe.insertNode(marker);
-			rect = marker.getBoundingClientRect();
-			marker.remove();
-			// Restore the original selection (anchor → focus) verbatim.
-			if (snap.anchorNode && snap.focusNode) {
-				try {
-					selection.setBaseAndExtent(
-						snap.anchorNode,
-						snap.anchorOffset,
-						snap.focusNode,
-						snap.focusOffset
-					);
-				} catch {
-					/* node may have been normalized away; ignore */
-				}
-			}
-			measuring = false;
-		}
-		if (!rect) return null;
-		return viewportDeltaToLocal(wrap, rect, lineHeight);
-	}
-
-	function measureCaret() {
-		if (!wrap || !editor || !caretTrailEnabled || !focused) return;
-		const measured = readCaretRect();
-		if (!measured) {
-			// Selection focus is outside the editor or unmeasurable: keep the
-			// caret where it is but stop drawing a trail.
-			clearTrail();
-			return;
-		}
-
-		caretH = measured.h;
-		if (customCaret) customCaret.style.height = `${caretH}px`;
-
-		if (!caretReady) {
-			// First placement — snap, no trail.
-			caretReady = true;
-			snapCaretTo(measured.x, measured.y);
-			return;
-		}
-
-		const dx = measured.x - visualX;
-		const dy = measured.y - visualY;
-		const dist = Math.hypot(dx, dy);
-		if (dist < 0.5) return; // no meaningful move
-
-		if (Math.abs(dy) > maxTrailVerticalJump()) {
-			// Real teleport (click far away, multi-line programmatic jump): snap
-			// without streaking a trail across the editor.
-			snapCaretTo(measured.x, measured.y);
-			return;
-		}
-
-		const typing = isTypingMove(dy);
-		trailOnly = typing;
-
-		if (typing) {
-			// Head snaps to the real caret immediately; only the trail smears.
-			originX = visualX;
-			originY = visualY;
-			setCaretVisual(measured.x, measured.y);
-		} else if (isLineCrossing(dy)) {
-			// Line wrap / newline: the literal old→new path would slash a diagonal
-			// across the editor. Instead, drop the comet in vertically — anchor the
-			// tail directly above the new caret position (same X, one line up) so
-			// the trail reads as a short vertical drop into the new line.
-			originX = measured.x;
-			originY = measured.y - dy;
-		} else {
-			// Same-line move: smear from the current rendered head toward the new target.
-			originX = visualX;
-			originY = visualY;
-		}
-		targetX = measured.x;
-		targetY = measured.y;
-		animStart = performance.now();
-		animating = true;
-		if (!raf) raf = requestAnimationFrame(animateCaret);
-	}
-
-	function setTrailQuad(
-		headX: number,
-		headY: number,
-		tailX: number,
-		tailY: number,
-		alpha: number
-	) {
-		if (!trailPoly) return;
-		// Build a quad spanning the bar caret from the tail position to the head
-		// position. The bar is `caretW` wide and `caretH` tall.
-		const x0 = headX;
-		const x1 = headX + caretW;
-		const tx0 = tailX;
-		const tx1 = tailX + caretW;
-		const pts = [
-			`${x0.toFixed(1)},${headY.toFixed(1)}`,
-			`${x1.toFixed(1)},${headY.toFixed(1)}`,
-			`${tx1.toFixed(1)},${(tailY + caretH).toFixed(1)}`,
-			`${tx0.toFixed(1)},${(tailY + caretH).toFixed(1)}`
-		];
-		trailPoly.setAttribute('points', pts.join(' '));
-		trailPoly.style.opacity = String(clampUnit(alpha) * baseTrailOpacity);
-	}
-
-	function animateCaret() {
-		if (!animating) {
-			raf = 0;
-			return;
-		}
-		const now = performance.now();
-		const duration = trailOnly ? typingTrailDuration() : moveDuration();
-		const progress = clampUnit((now - animStart) / duration);
-
-		let headX: number;
-		let headY: number;
-		let tailX: number;
-		let tailY: number;
-
-		if (trailOnly) {
-			// Head stays locked on target; tail smears from origin.
-			headX = targetX;
-			headY = targetY;
-			const tailEased = easeOutCirc(progress);
-			tailX = originX + (targetX - originX) * tailEased;
-			tailY = originY + (targetY - originY) * tailEased;
-			setCaretVisual(headX, headY);
-		} else {
-			// Head leads, tail follows with a delay so the smear stretches then
-			// collapses — same head/tail easing split as cursor_tail.glsl.
-			const headEased = easeOutCirc(progress);
-			const tailDelay = 0.18 + clampUnit(caretTrail.intensity) * 0.32;
-			const tailEased = easeOutCirc(clampUnit((progress - tailDelay) / (1 - tailDelay)));
-
-			headX = originX + (targetX - originX) * headEased;
-			headY = originY + (targetY - originY) * headEased;
-			tailX = originX + (targetX - originX) * tailEased;
-			tailY = originY + (targetY - originY) * tailEased;
-			setCaretVisual(headX, headY);
-		}
-
-		const span = Math.hypot(headX - tailX, headY - tailY);
-		if (span > 0.6) {
-			// Fade the trail out as the move completes.
-			setTrailQuad(headX, headY, tailX, tailY, 1 - progress * 0.35);
-		} else {
-			clearTrail();
-		}
-
-		if (progress >= 1) {
-			animating = false;
-			trailOnly = false;
-			originX = targetX;
-			originY = targetY;
-			snapCaretTo(targetX, targetY);
-			raf = 0;
-			return;
-		}
-		raf = requestAnimationFrame(animateCaret);
+	function scheduleCaretMeasure() {
+		scheduleCustomCaretMeasure(editor);
 	}
 
 	function resetCaretTrail() {
-		if (raf) cancelAnimationFrame(raf);
-		raf = 0;
-		caretReady = false;
-		animating = false;
-		trailOnly = false;
-		visualX = 0;
-		visualY = 0;
-		clearTrail();
+		resetCustomCaret(editor);
 	}
 
-	function scheduleCaretMeasure() {
-		if (!caretTrailEnabled || measuring) return;
-		requestAnimationFrame(measureCaret);
+	function onCaretStateChange(state: CustomCaretState) {
+		focused = state.focused;
+		caretReady = state.ready;
 	}
 
 	/** Build a non-editable inline chip element for a URL. */
@@ -669,7 +380,6 @@
 
 	function onInput() {
 		if (syncing || !editor) return;
-		lastInputAt = performance.now();
 		syncing = true;
 		decorateEditor();
 		syncing = false;
@@ -711,7 +421,6 @@
 		// Defer reset: some browsers fire the confirming keydown after compositionend.
 		setTimeout(() => {
 			composing = false;
-			lastInputAt = performance.now();
 			scheduleCaretMeasure();
 		}, 0);
 	}
@@ -737,16 +446,6 @@
 			e.preventDefault();
 			openLink(chip.dataset.url);
 		}
-	}
-
-	function onFocus() {
-		focused = true;
-		scheduleCaretMeasure();
-	}
-
-	function onBlur() {
-		focused = false;
-		resetCaretTrail();
 	}
 
 	function insertPlainText(text: string) {
@@ -901,39 +600,17 @@
 	});
 
 	$effect(() => {
-		if (!caretTrailEnabled) {
-			resetCaretTrail();
-			return;
-		}
-		const onSelectionChange = () => {
-			scheduleCaretMeasure();
-			updateMentionState();
-		};
-		const onResize = () => scheduleCaretMeasure();
+		const onSelectionChange = () => updateMentionState();
 		document.addEventListener('selectionchange', onSelectionChange);
-		window.addEventListener('resize', onResize);
-		scheduleCaretMeasure();
 		return () => {
 			document.removeEventListener('selectionchange', onSelectionChange);
-			window.removeEventListener('resize', onResize);
-			resetCaretTrail();
 		};
 	});
-
-	$effect(() => {
-		const el = editor;
-		if (!el || !caretTrailEnabled) return;
-		const onScroll = () => scheduleCaretMeasure();
-		el.addEventListener('scroll', onScroll, { passive: true });
-		return () => el.removeEventListener('scroll', onScroll);
-	});
-
-	onDestroy(resetCaretTrail);
 
 	let isEmpty = $derived(value.trim() === '');
 </script>
 
-<div bind:this={wrap} class="rce-wrap" style={caretStyle}>
+<div bind:this={wrap} class="rce-wrap">
 	{#if isEmpty}
 		<div class="rce-placeholder" aria-hidden="true">{placeholder}</div>
 	{/if}
@@ -942,13 +619,21 @@
 			<svg class="rce-trail" focusable="false">
 				<polygon bind:this={trailPoly}></polygon>
 			</svg>
-			<span bind:this={customCaret} class="rce-caret"></span>
+			<span bind:this={caretEl} class="rce-caret"></span>
 		</div>
 	{/if}
 	<div
 		bind:this={editor}
 		class="rce-editor scrollbar-none"
 		class:trail-enabled={caretTrailEnabled}
+		use:customCaret={{
+			wrap,
+			caret: caretEl,
+			trail: trailPoly,
+			caretTrail,
+			color: caretColor,
+			onStateChange: onCaretStateChange
+		}}
 		contenteditable="true"
 		role="textbox"
 		tabindex="0"
@@ -960,8 +645,8 @@
 		oncompositionstart={onCompositionStart}
 		oncompositionend={onCompositionEnd}
 		onclick={onEditorClick}
-		onfocus={onFocus}
-		onblur={onBlur}
+		onfocus={updateMentionState}
+		onblur={updateMentionState}
 	></div>
 </div>
